pyproject.toml
[project]
name = "billmunshi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.40.5",
    "celery[redis]>=5.5.3",
    "celery-progress>=0.5",
    "django>=5.2.4",
    "django-allauth[mfa,socialaccount]>=65.10.0",
    "django-anymail[mailgun]>=13.0.1",
    "django-environ>=0.12.0",
    "django-health-check>=3.20.0",
    "django-waffle>=5.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-api-key>=3.1.0",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "redis>=5.2.1",
    "pytz>=2025.2",
    "user-agents>=2.2.0",
    "django-jsonfield>=1.4.1",
    "django-cors-headers>=4.7.0",
    "drf-nested-routers>=0.94.2",
]
